Group elements

GroupsCore defines abstract type GroupElement, which all implementations of group elements should subtype.

Obligatory methods

The elements which are not isbits should extend

Base.deepcopy_internal(g::GroupElement, ::IdDict)

according to Base.deepcopy docstring. Due to our assumption on parents of group elements (acting as local singleton objects), a group element and its deepcopy should have identical (i.e. ===) parents.

The remaining obligatory methods are:

Base.parent — Method

parent(g::GroupElement)

Return the parent object of the group element.

AbstractAlgebra.Generic.parent_type — Method

parent_type(::Type{<:GroupElement})

Return the type of parent object, based on the type of its elements.

Base.:== — Method

==(g::GEl, h::GEl) where {GEl <: GroupElement}

Return true if and only if the mathematical equality $g = h$ holds.

Note

This function may not return, due to unsolvable word problem.

GroupsCore.isfiniteorder — Method

isfiniteorder(g::GroupElement)

Return true if $g$ is of finite order, possibly without computing it.

Base.inv — Method

inv(g::GroupElement)

Return $g^{-1}$, the group inverse.

Base.:* — Method

*(g::GEl, h::GEl) where {GEl <: GroupElement}

Return $g h$, the result of group binary operation.

Note

If finiteness of a group can be decided based on its type there is no need to extend isfiniteorder.

Implemented methods

Using the obligatory methods we implement the rest of the functions in GroupsCore. For starters, the first of these are:

:(^)(::GroupElement, ::Integer)
:(/)(::GEl, ::GEl) where {GEl <: GroupElement}

and

Base.one — Method

one(g::GroupElement)

Return the identity element in the group of $g$.

Base.isequal — Method

isequal(g::GEl, h::GEl) where {GEl <: GroupElement}

Return the "best effort" equality for group elements.

The implication isequal(g, h) → $g = h$, must be always satisfied, i.e. isequal(g, h) might return false even if g == h holds (i.e. $g$ and $h$ are mathematically equal).

For example, in a finitely presented group, isequal may return the equality of words.

AbstractAlgebra.Generic.order — Method

order(::Type{I} = BigInt, g::GroupElement) where {I <: Integer}

Return the order of $g$ as an instance of I. If $g$ is of infinite order GroupsCore.InfiniteOrder exception will be thrown.

Warning

Only arbitrary sized integers are required to return a mathematically correct answer.

Base.conj — Function

conj(g::GEl, h::GEl) where {GEl <: GroupElement}

Return conjugation of $g$ by $h$, i.e. $h^{-1} g h$.

Base.:^ — Method

^(g::GEl, h::GEl) where {GEl <: GroupElement}

Alias for conj.

GroupsCore.commutator — Function

commutator(g::GEl, h::GEl, k::GEl...) where {GEl <: GroupElement}

Return the left associative iterated commutator $[[g, h], ...]$, where $[g, h] = g^{-1} h^{-1} g h$.

Performance modifications

Some of the mentioned implemented methods may be altered for performance reasons:

isequal

Base.:(^)(g::GroupElement, n::Integer)

order(::Type{<:Integer}, g::GroupElement)

Base.hash(::GroupElement, ::UInt)

Base.similar — Method

similar(g::GroupElement)

Return a group element sharing the parent with $g$. Might be arbitrary and possibly uninitialized.

Base.isone — Method

isone(g::GroupElement)

Return true if $g$ is the identity element.

Mutable API

Warning

Work-in-progress.

Mutable API is considered private and hence may change between versions without warning.

For the purpose of mutable arithmetic the following methods may be overloaded to provide more tailored versions for a given type and reduce the allocations. These functions should be used when writing libraries, in performance critical sections. However one should only use the returned value and there are no guarantees on in-place modifications actually happening.

All of these functions (possibly) alter only the first argument, and must unalias their arguments when necessary.

GroupsCore.one! — Function

one!(g::GroupElement)

Return one(g), possibly modifying g.

AbstractAlgebra.Generic.inv! — Function

inv!(out::GEl, g::GEl) where {GEl <: GroupElement}

Return inv(g), possibly modifying out. Aliasing of g with out is allowed.

AbstractAlgebra.Generic.mul! — Function

mul!(out::GEl, g::GEl, h::GEl) where {GEl <: GroupElement}

Return $g h$, possibly modifying out. Aliasing of g or h with out is allowed.

GroupsCore.div_left! — Function

div_left!(out::GEl, g::GEl, h::GEl) where {GEl <: GroupElement}

Return $h^{-1} g$, possibly modifying out. Aliasing of g or h with out is allowed.

GroupsCore.div_right! — Function

div_right!(out::GEl, g::GEl, h::GEl) where {GEl <: GroupElement}

Return $g h^{-1}$, possibly modifying out. Aliasing of g or h with out is allowed.

GroupsCore.conj! — Function

conj!(out::GEl, g::GEl, h::GEl) where {GEl <: GroupElement}

Return $h^{-1} g h$, possibly modifying out. Aliasing of g or h with out is allowed.

GroupsCore.commutator! — Function

commutator!(out::GEl, g::GEl, h::GEl) where {GEl <: GroupElement}

Return $g^{-1} h^{-1} g h$, possibly modifying out. Aliasing of g or h with out is allowed.